data/pyproject.toml
[project]
name = "prismatics-data"
version = "0.1.0"
description = "Data Project for Prismatics"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.9,<3.14"
dependencies = [
    "prefect>=3.4.10",
    "python-dotenv>=1.1.1",
    "polars>=1.32.0",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.42",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.16.4",
]

[tool.pdm]
distribution = false

[tool.pdm.build]
package-dir = "src"

[tool.pdm.scripts]
prefect = "prefect server start"
pipeline = { cmd = "python -m main", env = { PYTHONPATH = "src" } }
